+++ vm/smx-v1-image.h
@@ -236,7 +236,7 @@ class SmxV1Image
   template <typename SymbolType, typename DimType>
   const char* getFunctionName(const SymbolType* syms, const char** filename, uint32_t index) const;
   template <typename SymbolType, typename DimType>
-  bool getFunctionAddress(const SymbolType* syms, const char* function, ucell_t* funcaddr, uint32_t index) const;
+  bool getFunctionAddress(const SymbolType* syms, const char* function, ucell_t* funcaddr, size_t index) const;
 
 
   const smx_rtti_table_header* findRttiSection(const char* name) const {
